rename: 피드백 1차 반영 카드 컴포넌트 > 공통 컴포넌트 제외
